Move file upload test to integration test. Reference: #5320
